+++ yash-env/src/function.rs
@@ -18,18 +18,53 @@
 //!
 //! This module provides data types for defining shell functions.
 
+use crate::Env;
+use crate::semantics::ExitStatus;
 use std::borrow::Borrow;
 use std::collections::HashSet;
+use std::fmt::Debug;
+use std::fmt::Display;
 use std::hash::Hash;
 use std::hash::Hasher;
 use std::iter::FusedIterator;
+use std::pin::Pin;
 use std::rc::Rc;
 use thiserror::Error;
 use yash_syntax::source::Location;
 use yash_syntax::syntax::FullCompoundCommand;
 
+/// Trait for the body of a [`Function`]
+pub trait FunctionBody: Debug + Display {
+    /// Executes the function body in the given environment.
+    #[allow(async_fn_in_trait)] // We don't support Send
+    async fn execute(&self, env: &mut Env) -> crate::semantics::Result<ExitStatus>;
+}
+
+/// Dyn-compatible adapter for the [`FunctionBody`] trait
+///
+/// This is a dyn-compatible version of the [`FunctionBody`] trait.
+///
+/// This trait is automatically implemented for all types that implement
+/// [`FunctionBody`].
+pub trait FunctionBodyObject: Debug + Display {
+    /// Executes the function body in the given environment.
+    fn execute<'a>(
+        &'a self,
+        env: &'a mut Env,
+    ) -> Pin<Box<dyn Future<Output = crate::semantics::Result<ExitStatus>> + 'a>>;
+}
+
+impl<T: FunctionBody + ?Sized> FunctionBodyObject for T {
+    fn execute<'a>(
+        &'a self,
+        env: &'a mut Env,
+    ) -> Pin<Box<dyn Future<Output = crate::semantics::Result<ExitStatus>> + 'a>> {
+        Box::pin(self.execute(env))
+    }
+}
+
 /// Definition of a function.
-#[derive(Clone, Debug, Eq, PartialEq)]
+#[derive(Clone, Debug)]
 pub struct Function {
     /// String that identifies the function.
     pub name: String,
@@ -89,6 +124,21 @@ impl Function {
         self.read_only_location.is_some()
     }
 }
+
+/// Compares two functions for equality.
+///
+/// Two functions are considered equal if all their members are equal.
+/// This includes comparing the `body` members by pointer equality.
+impl PartialEq for Function {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+            && Rc::ptr_eq(&self.body, &other.body)
+            && self.origin == other.origin
+            && self.read_only_location == other.read_only_location
+    }
+}
+
+impl Eq for Function {}
 
 /// Wrapper of [`Function`] for inserting into a hash set.
 ///