+++ yash-env/CHANGELOG.md
@@ -13,9 +13,17 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 ### Added
 
+- `function::FunctionBody`: Trait for the body of a shell function.
+- `function::FunctionBodyObject`: Dyn-compatible adapter for the `FunctionBody`
+  trait.
 - `parser::IsName`: Wrapper for a function that checks if a string is a valid
   variable name. This allows modules to check variable names without directly
   depending on the `yash-syntax` crate.
+
+### Changed
+
+- `impl PartialEq for function::Function`: Now compares the function bodies
+  using pointer equality instead of deep equality.
 
 ### Removed